Refactor: jmarkdown <script> extension
The extension which looks for <script> elements with data-type of 'jmarkdown' or 'jmarkdown-postprocess' has been refactored into script-blocks.js.

In order to make this work, I had to create a utility module which created a single instance of runInThisContext and exported it, so that it could be shared by both the main module (from index.js) and script-blocks.js.