Code cleanup: 'marked' object
In refactoring the code, it turns out that several of the files were relying on 'marked' being resolved by implicitly relying on the fact that 'marked' was declared in the global scope by index.js.  I've moved the creation of the global 'marked' object into utils.js, which then exposes it via export.