Refactor: processing metadata header
This is a pretty big refactoring: the function processYAMLheader() and all of the helper functions it invokes were moved into metadata-header.js.  This made index.js less cluttered.

In addition, the 'metadata' variable is no longer the default export from metadata-header.js.  Since that file had to export a few other functions and variables, there was no point in having both a default export and named exports.

Also - I added the function registerExtension() to utils.js, which makes registering extensions defined in the metadata header easier.

Yes - this is probably too much for a single commit. Sorry. ;-)