Refactor: strategic-form games
I moved the code for strategic-form games into a separate file. In addition, I rewrote part of it to take advantage of the fact that my extended-directives.js extension allows the tokeniser to be specified, in addition to the renderer. This allowed me to add support for running the strategy labels, as well as the row, column and caption labels, through the tokeniser, providing support for typesetting and even footnotes!