Added 'failsafe' option to markdown-demo
Since I can't fix the bug which causes the parsed output in the right panel from occasionally being wrong, I added this option which - so far - seems to always work.  Passing the 'failsafe' attribute to markdown-demo uses a separate, independent copy of the marked parser (which should be configured exactly the same) to generate the HTML for the right panel. So far this seems to work.