Added basic project initialisation
If 'jmarkdown init [filename]' is typed on the command line, it will check to see if .jmarkdown/ and .jmarkdown/config.json exist in the current directory. If not, the directory will be created and a config-template.json will be created as well. If the optional filename is given, a barebones file of that type will be created.