Embedded JavaScript now respects inline punctuation
acorn.js would frequently throw SyntaxErrors if the inline JavaScript was surrounded by punctuation. This was often caused because acorn.js would try to parse just a little too far and didn't know when to fall back to a valid substring. This commit fixes that, I think. It's a bit ugly and would benefit from refactoring to remove redundant code, but that's for a later time.