Add print utility files during init
When the user invokes 'jmarkdown init' there is now a new option '-p --print' that says to copy over a package.json file and print.js file, which can be used to generate PDF output from a jmarkdown-produced HTML file using puppeteer. This ensures that all javascript is processed so the bibliographies appear.